(feat) : UI tweaks to consolidate Invoice actions
